Refactor functions to files, default progress logs
* Update README
* New patset.py
